# httprunner 升级到3.1.11 支持项目，模块，用例的配置参数

+++ autotest/services/api_services/env.py
@@ -1,16 +1,18 @@
-import traceback
-from typing import Dict, Any
+from typing import Dict, Any, Text
 
-from loguru import logger
-
-from autotest.serialize.api_serializes.env import EnvQuerySchema, EnvListSchema, EnvSaveOrUpdateSchema
 from autotest.models.api_models import Env
+from autotest.serialize.api_serializes.env import EnvQuerySchema, EnvListSchema, EnvSaveOrUpdateSchema
 from autotest.utils.api import parse_pagination
 
 
 class EnvService:
     @staticmethod
-    def list(**kwargs: Any) -> Dict:
+    def list(**kwargs: Any) -> Dict[Text, Any]:
+        """
+        获取环境列表
+        :param kwargs:
+        :return:
+        """
         parsed_data = EnvQuerySchema().load(kwargs)
         data = parse_pagination(Env.get_list(**parsed_data))
         _result, pagination = data.get('result'), data.get('pagination')
@@ -21,7 +23,12 @@ class EnvService:
         return result
 
     @staticmethod
-    def save_or_update(**kwargs: Any) -> Env:
+    def save_or_update(**kwargs: Any) -> "Env":
+        """
+        保存更新环境
+        :param kwargs:
+        :return:
+        """
         parsed_data = EnvSaveOrUpdateSchema().load(kwargs)
         env_id = kwargs.get('id', None)
         env_info = Env.get(env_id) if env_id else Env()
@@ -30,9 +37,10 @@ class EnvService:
 
     @staticmethod
     def deleted(id: int):
-        """删除项目"""
-        try:
-            env = Env.get(id)
-            env.delete()
-        except Exception as err:
-            logger.error(traceback.format_exc())
+        """
+        删除环境
+        :param id:
+        :return:
+        """
+        env = Env.get(id)
+        env.delete() if env else ...
